+++ src/main/java/com/bd2FarmaUNLa/farmaUNLa/model/ObraSocial.java
@@ -11,21 +11,21 @@ public class ObraSocial {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private int id;
+	private long id;
 	private String nombre;
 	
 	public ObraSocial(){} 
 	
-	public ObraSocial(int id, String nombre) {
+	public ObraSocial(long id, String nombre) {
 		this.id = id;
 		this.nombre = nombre;
 	}
 	
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
